Minor optimizations; Disable fast forwarding in certain modes.

+++ modules/io/src/main/java/net/multiphasicapps/io/InflaterInputStream.java
@@ -33,28 +33,28 @@ public class InflaterInputStream
 		32768;
 	
 	/** No compression. */
-	static final int _TYPE_NO_COMPRESSION =
+	static final byte _TYPE_NO_COMPRESSION =
 		0b00;
 	
 	/** Fixed huffman table compression. */
-	static final int _TYPE_FIXED_HUFFMAN =
+	static final byte _TYPE_FIXED_HUFFMAN =
 		0b01;
 	
 	/** Dynamic huffman table compression. */
-	static final int _TYPE_DYNAMIC_HUFFMAN =
+	static final byte _TYPE_DYNAMIC_HUFFMAN =
 		0b10;
 	
 	/** An error. */
-	static final int _TYPE_ERROR =
+	static final byte _TYPE_ERROR =
 		0b11;
 	
 	/** The maximum number of bits in the code length tree. */
-	private static final int _MAX_BITS =
+	private static final byte _MAX_BITS =
 		15;
 	
 	/** Shuffled bit values when reading values. */
-	private static final int[] _SHUFFLE_BITS =
-		new int[]
+	private static final byte[] _SHUFFLE_BITS =
+		new byte[]
 		{
 			16, 17, 18, 0, 8, 7, 9, 6, 10, 5, 11, 4, 12, 3, 13, 2, 14, 1, 15
 		};
@@ -563,7 +563,7 @@ public class InflaterInputStream
 		
 		// Read lengths, they are just 3 bits but their placement values are
 		// shuffled since some sequences are more common than others
-		int[] hsbits = InflaterInputStream._SHUFFLE_BITS;
+		byte[] hsbits = InflaterInputStream._SHUFFLE_BITS;
 		for (int next = 0; next < __dhclen; next++)
 			rawcodelens[hsbits[next]] = this.__readBits(3, false);
